Moved the individual file data stores to their own package
